third_party/blink/web_tests/platform/generic/fast/events/event-function-toString-expected.txt

Test serialization of event handlers for bug 25544.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS normalizedFunctionString(div.onclick.toString()) is "function onclick(event) {test();}"
PASS normalizedFunctionString(path.onclick.toString()) is "function onclick(evt) {test();}"
PASS successfullyParsed is true

TEST COMPLETE
